backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
anthropic>=0.28.0
httpx>=0.27.0
tweepy>=4.14.0
apscheduler>=3.10.0
google-cloud-secret-manager>=2.20.0
